Break individual classes out into a directory. (and fix embed test)
