+++ lib/dc/embed/search.rb
@@ -1,0 +1,6 @@
+module DC
+  module Embed
+    class Search < Base
+    end
+  end
+end
